// src/api/controller/v1/controller_pb.d.ts
export class RunCrawlRequest extends jspb.Message {
  getJobId(): string;
  setJobId(value: string): RunCrawlRequest;

  getSeedId(): string;
  setSeedId(value: string): RunCrawlRequest;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): RunCrawlRequest.AsObject;
  static toObject(includeInstance: boolean, msg: RunCrawlRequest): RunCrawlRequest.AsObject;
  static serializeBinaryToWriter(message: RunCrawlRequest, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): RunCrawlRequest;
  static deserializeBinaryFromReader(message: RunCrawlRequest, reader: jspb.BinaryReader): RunCrawlRequest;
}

export namespace RunCrawlRequest {
  export type AsObject = {
    jobId: string,
    seedId: string,
  }
}

export class RunCrawlReply extends jspb.Message {
  getJobExecutionId(): string;
  setJobExecutionId(value: string): RunCrawlReply;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): RunCrawlReply.AsObject;
  static toObject(includeInstance: boolean, msg: RunCrawlReply): RunCrawlReply.AsObject;
  static serializeBinaryToWriter(message: RunCrawlReply, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): RunCrawlReply;
  static deserializeBinaryFromReader(message: RunCrawlReply, reader: jspb.BinaryReader): RunCrawlReply;
}

export namespace RunCrawlReply {
  export type AsObject = {
    jobExecutionId: string,
  }
}
// ...
export class CrawlerStatus extends jspb.Message {
  getRunstatus(): RunStatus;
  setRunstatus(value: RunStatus): CrawlerStatus;

  getBusycrawlhostgroupcount(): number;
  setBusycrawlhostgroupcount(value: number): CrawlerStatus;

  getQueuesize(): number;
  setQueuesize(value: number): CrawlerStatus;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CrawlerStatus.AsObject;
  static toObject(includeInstance: boolean, msg: CrawlerStatus): CrawlerStatus.AsObject;
  static serializeBinaryToWriter(message: CrawlerStatus, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CrawlerStatus;
  static deserializeBinaryFromReader(message: CrawlerStatus, reader: jspb.BinaryReader): CrawlerStatus;
}

export namespace CrawlerStatus {
  export type AsObject = {
    runstatus: RunStatus,
    busycrawlhostgroupcount: number,
    queuesize: number,
  }
}

export enum RunStatus {
  RUNNING = 0,
  PAUSED = 1,
  PAUSE_REQUESTED = 2,
}